moved images into separate folder, updated all image imports, added stylized photos
