+++ .vscode/settings.json
@@ -1,0 +1,7 @@
+{
+    "workbench.colorCustomizations": {
+        "activityBar.background": "#0A2F52",
+        "titleBar.activeBackground": "#0E4173",
+        "titleBar.activeForeground": "#FAFCFE"
+    }
+}
